requirements.txt
pandas
rpi.gpio
PyMLX90614
RPi.bme280
paho-mqtt